requirements.txt
chromadb
cohere
datasets
dspy-ai==2.4.9
easydict
jupyter
llama-index==0.10.35
llama-index-agent-openai==0.2.4
llama-index-cli==0.1.12
llama-index-core==0.10.35
llama-index-embeddings-azure-openai==0.1.9
llama-index-embeddings-huggingface==0.1.5
llama-index-embeddings-openai==0.1.9
llama-index-embeddings-text-embeddings-inference==0.1.2
llama-index-graph-stores-nebula==0.1.2
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-anthropic==0.1.11
llama-index-llms-azure-openai==0.1.7
llama-index-llms-cohere==0.2.0
llama-index-llms-huggingface==0.2.3
llama-index-llms-huggingface-api==0.1.0
llama-index-llms-openai==0.1.18
llama-index-llms-openai-like
llama-index-llms-openai-like==0.1.3
llama-index-llms-vllm==0.1.8
llama-index-multi-modal-llms-openai==0.1.5
llama-index-postprocessor-cohere-rerank==0.1.4
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.22
llama-index-readers-llama-parse==0.1.4
llama-index-readers-wikipedia==0.1.4
llama-index-vector-stores-chroma==0.1.8
openpyxl
pandas
pyparsing
python-dotenv
pyyaml
scikit-learn
seaborn
statsmodels
tonic-validate==6.1.0
